move to namespace and matrix-free folder
